take advantage of simpler next v9
