+++ .gitignore
@@ -42,3 +42,5 @@ docs/
 
 # =====================================
 # CUSTOM
+
+.dev
