+++ client/components/events.tsx
@@ -10,7 +10,7 @@ import {
 import { useInterval, useMetaKey } from '@bevry/hooks'
 import Daet from 'daet'
 
-import { RichEventsType, RichEventType } from '../../client/events'
+import { RichEventsType, RichEventType } from '../../shared/types'
 import { podcastJoinUrl, podcastWatchUrl } from '../../shared/config'
 
 // Event
