+++ now.json
@@ -2,217 +2,7 @@
   "version": 2,
   "name": "bevry-meet",
   "alias": "meet.bevry.me",
-  "routes": [
-    {
-      "src": "/api/(?<key>[^./]*)/?",
-      "dest": "/server/api/$key.ts",
-      "headers": {
-        "Access-Control-Allow-Origin": "*"
-      }
-    },
-    {
-      "src": "/calendar.ics/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/calendar/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/discord/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discord.gg/j2PqA7S"
-      }
-    },
-    {
-      "src": "/facebook-liberal-group/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.facebook.com/groups/1408546319260687/"
-      }
-    },
-    {
-      "src": "/facebook-study-group/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.facebook.com/groups/198567587146349/"
-      }
-    },
-    {
-      "src": "/favicon.ico",
-      "dest": "/static/favicon.ico"
-    },
-    {
-      "src": "/feedback/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/c/feedback"
-      }
-    },
-    {
-      "src": "/forum/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community"
-      }
-    },
-    {
-      "src": "/google-calendar-support/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/home/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://jordanbpeterson.community"
-      }
-    },
-    {
-      "src": "/ios-calendar-support/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/lecture-notes/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/search?q=tags%3Ajbp%2Blecture"
-      }
-    },
-    {
-      "src": "/macos-calendar-support/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/meetups/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.meetup.com/find/?allMeetups=false&keywords=jordan+peterson&radius=Infinity"
-      }
-    },
-    {
-      "src": "/newsletter/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://goo.gl/forms/qgIjkJ8SYEpLCVY83"
-      }
-    },
-    {
-      "src": "/outlook-calendar-support/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/podcast-join/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://jordanbpeterson.community"
-      }
-    },
-    {
-      "src": "/podcast/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.youtube.com/channel/UCPkobzBsAIjpItonUT7AU9Q"
-      }
-    },
-    {
-      "src": "/reading-group/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.goodreads.com/group/show/253798-jordan-b-peterson-reading-group"
-      }
-    },
-    {
-      "src": "/reading-list/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.goodreads.com/group/bookshelf/253798-jordan-b-peterson-reading-group?utf8=✓&order=d&sort=date_finished&view=main&per_page=200"
-      }
-    },
-    {
-      "src": "/reddit/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.reddit.com/r/JordanPeterson"
-      }
-    },
-    {
-      "src": "/robots.txt",
-      "dest": "/static/robots.txt"
-    },
-    {
-      "src": "/source/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://github.com/bevry/jordanbpeterson.community"
-      }
-    },
-    {
-      "src": "/study-group-calendar/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/please-unsubscribe-from-the-calendar-subscription/474"
-      }
-    },
-    {
-      "src": "/study-group-hangout/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/about-the-meetings/92"
-      }
-    },
-    {
-      "src": "/study-group-recordings/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.youtube.com/channel/UCPkobzBsAIjpItonUT7AU9Q"
-      }
-    },
-    {
-      "src": "/study-group/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://discuss.jordanbpeterson.community/t/about-the-meetings/92"
-      }
-    },
-    {
-      "src": "/twitter/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://twitter.com/JBPCommunity"
-      }
-    },
-    {
-      "src": "/youtube/?",
-      "status": 302,
-      "headers": {
-        "Location": "https://www.youtube.com/channel/UCPkobzBsAIjpItonUT7AU9Q"
-      }
-    }
-  ],
   "builds": [
-    {
-      "src": "server/api/*",
-      "use": "@now/node@canary"
-    },
     {
       "src": "package.json",
       "use": "@now/next"
