--- server/api/google-oauth.ts
@@ -1,34 +1,0 @@
-import { Http2ServerRequest, Http2ServerResponse } from 'http2'
-import { createEnv } from '../lib/env'
-import { setKeyValue } from '../lib/fauna'
-import { getOAuthLoginUrl, getOAuth } from '../lib/google'
-import { URL } from 'url'
-
-function processUserAuth(code: string): Promise<ReturnType<typeof getOAuth>> {
-	const client = getOAuth()
-	return client.getToken(code).then(function({ tokens }) {
-		client.setCredentials(tokens)
-		return setKeyValue('google-oauth-result', createEnv(tokens)).then(
-			() => client
-		)
-	})
-}
-
-export default async function sendEvents(
-	req: Http2ServerRequest,
-	res: Http2ServerResponse
-) {
-	const url = new URL(req.url)
-	const code = url.searchParams.get('code')
-	if (code) {
-		processUserAuth(code)
-		res.end('ok')
-	} else {
-		const authUrl = getOAuthLoginUrl()
-		res.writeHead(302, {
-			Location: authUrl,
-			'Content-Type': 'text/plain'
-		})
-		res.end(`redirecting to ${authUrl}`)
-	}
-}
