+++ shared/config.ts
@@ -1,10 +1,9 @@
 export const DEVELOPMENT = process.env.NODE_ENV === 'development'
-export const eventsUrl = 'https://meet.bevry.me/api/events/'
-
+export const eventsUrl = DEVELOPMENT
+	? 'http://localhost:3000/api/events'
+	: '/api/events'
 export const podcastJoinUrl = 'https://meet.bevry.me/join'
-
 export const podcastWatchUrl = 'https://meet.bevry.me/watch'
-
 export const expiresValue = 3
 export const expiresUnit = 'hour'
 export const googleCalendarId =
